requirements.txt
python-dateutil==2.8.2
selenium==4.10.0
webdriver-manager==3.8.6

